requirements.txt
Flask==2.0.3
Flask_Cors==3.0.10
GDAL==3.3.1
labelme==3.18.0
matplotlib==3.5.3
networkx==2.8.4
numpy==1.23.1
opencv_python==4.6.0.66
osgeo==0.0.1
pandas==1.4.1
Pillow==9.4.0
scikit_image==0.19.3
scikit_learn==1.2.0
scipy==1.8.0
skimage==0.0
torch==1.11.0+cu113
torch_summary==1.4.4
torchsummary==1.5.1
torchvision==0.12.0+cu113
tornado==6.2
tqdm==4.64.0
